fix(layer): layer edition is unified, but missing capabilities
